♻️ refactor: useState jotai로 수정
