+++ apps/web/src/components/Header/_components/ModeSwitchButton/index.ts
@@ -1,2 +1,1 @@
 export * from './ModeSwitchButton';
-export * from './UserButton';
