+++ apps/web/src/components/Header/_components/UserButton/index.ts
@@ -1,0 +1,1 @@
+export * from './UserButton';
